Polishing

See merge request !4

+++ Vagrantfile
@@ -4,6 +4,9 @@
 # Make sure essential plugins are installed
 require File.dirname(__FILE__)+"/scripts/dependency_manager"
 require File.dirname(__FILE__)+"/scripts/update_apps"
+require File.dirname(__FILE__)+"/scripts/utilities"
+require 'fileutils'
+
 # If plugins have been installed, rerun the original vagrant command and abandon this one
 if not check_plugins ["vagrant-cachier", "vagrant-triggers", "vagrant-reload"]
   exec "vagrant #{ARGV.join(' ')}" unless ARGV[0] == 'plugin'
@@ -18,40 +21,57 @@ Vagrant.configure(2) do |config|
   config.vm.box_version      = "0.3.0"
   config.vm.box_check_update = false
   config.ssh.forward_agent = true
+  
+  # Configure cached packages to be shared between instances of the same base box.
+ 	# More info on http://fgrehm.viewdocs.io/vagrant-cachier/usage
+  config.cache.scope       = :box
+  # Make cachier only cache yum instead of rooting about trying to figure things out itself
+  config.cache.auto_detect = false
+  config.cache.enable :yum
 
   #Only if vagrant up/reload/resume do want to create dev-env configuration
   if ['up', 'reload', 'resume'].include? ARGV[0]
     # Check if a DEV_ENV_CONTEXT_FILE exists, to prevent prompting for dev-env configuration choice on each vagrant up
     if File.exists?(DEV_ENV_CONTEXT_FILE)
-      print "This dev env has been provisioned to run for the repo: #{File.read(DEV_ENV_CONTEXT_FILE)}\n"
+      puts colorize_lightblue("This dev env has been provisioned to run for the repo: #{File.read(DEV_ENV_CONTEXT_FILE)}")
      else
-      print "This is a universal dev env.\n"
-      print "Please enter the url of your dev env repo:"
+      puts colorize_lightblue("This is a universal dev env.")
+      print colorize_yellow("Please enter the url of your dev env repo (SSH): ")
       app_grouping = STDIN.gets.chomp
       File.open(DEV_ENV_CONTEXT_FILE, "w+") { |file| file.write(app_grouping) }
     end
 
     #Check if dev-env-project exists, and if so pull the dev-env configuration. Otherwise clone it.
-    if Dir.exists?("dev-env-project")
-      command_successful = system 'git', '-C', 'dev-env-project', 'pull'
+    puts colorize_lightblue("Retrieving custom configuration repo files:")
+    if Dir.exists?(File.dirname(__FILE__) + '/dev-env-project')
+      command_successful = system 'git', '-C', File.dirname(__FILE__) + '/dev-env-project', 'pull'
     else
-      command_successful = system 'git', 'clone', File.read(DEV_ENV_CONTEXT_FILE), 'dev-env-project'
+      command_successful = system 'git', 'clone', File.read(DEV_ENV_CONTEXT_FILE), File.dirname(__FILE__) + '/dev-env-project'
     end
 
     #Error if git clone or pulling failed
     if command_successful == false
-      puts "Something went wrong when cloning/pulling the dev-env configuration project"
+      puts colorize_red("Something went wrong when cloning/pulling the dev-env configuration project")
       exit 1
     end
 
     #Call the ruby function to pull/clone all the apps found in dev-env-project/configuration.yml
-    update_apps()
+    puts colorize_lightblue("Updating apps:")
+    update_apps(File.dirname(__FILE__))
   end
 
   # In the event of user requesting a vagrant destroy, remove DEV_ENV_CONTEXT_FILE created on provisioning
   config.trigger.before :destroy do
-    print "Dumping the DEV_ENV_CONTEXT_FILE before destroying the VM...\n"
-    File.delete(DEV_ENV_CONTEXT_FILE)
+    confirm = nil
+    until ["Y", "y", "N", "n"].include?(confirm)
+      confirm = ask colorize_yellow("Would you like to keep your custom dev-env configuration files? (Y/N) ")
+    end
+    if confirm.upcase == "N"
+      File.delete(DEV_ENV_CONTEXT_FILE)
+      if Dir.exists?(File.dirname(__FILE__) + '/dev-env-project')
+        FileUtils.rm_r File.dirname(__FILE__) + '/dev-env-project'
+      end
+    end
   end
 
   # Run script to configure environment
